LT-20408: Allow adding valid characters with diacritics
* Cast the return type of Normalizer2.GetCombiningClass to byte.
This will have to do until we can upgrade to the latest version
of ICU.NET, which has the proper return type.

+++ src/SIL.LCModel.Core/Text/CustomIcu.cs
@@ -357,7 +357,7 @@ namespace SIL.LCModel.Core.Text
 		public static UcdProperty GetCombiningClassInfo(int characterCode)
 		{
 			var normalizer = GetIcuNormalizer(FwNormalizationMode.knmNFC);
-			return UcdProperty.GetInstance(normalizer.GetCombiningClass(characterCode));
+			return UcdProperty.GetInstance((byte)normalizer.GetCombiningClass(characterCode));
 		}
 
 		/// ------------------------------------------------------------------------------------
